fix rebase after otel restructure
